added auditor and query

+++ Repository/IGroupRepository.cs
@@ -6,9 +6,9 @@ namespace RestApi.Repository
     public interface IGroupRepository
     {
         List<Group> GetAll();
-        Group GetOne(long id);
+        Group GetOne(short id);
         Group Save(Group group);
         Group Update(Group group);
-        void Delete(long id);
+        void Delete(short id);
     }
 }
